Added int64 support to the script engine
Added support for signed 64 bit integer value computation into the script engine.
This is required because newer official features require support for bigger numbers inside the scripts.

This also cleans up a lot of messy code and fixes a few issues like script stack corruptions.

Thanks to @aleos89 and everyone who supported me with this.

+++ sql-files/main.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS `acc_reg_num` (
   `account_id` int(11) unsigned NOT NULL default '0',
   `key` varchar(32) binary NOT NULL default '',
   `index` int(11) unsigned NOT NULL default '0',
-  `value` int(11) NOT NULL default '0',
+  `value` bigint(11) NOT NULL default '0',
   PRIMARY KEY (`account_id`,`key`,`index`),
   KEY `account_id` (`account_id`)
 ) ENGINE=MyISAM;
@@ -274,7 +274,7 @@ CREATE TABLE IF NOT EXISTS `char_reg_num` (
   `char_id` int(11) unsigned NOT NULL default '0',
   `key` varchar(32) binary NOT NULL default '',
   `index` int(11) unsigned NOT NULL default '0',
-  `value` int(11) NOT NULL default '0',
+  `value` bigint(11) NOT NULL default '0',
   PRIMARY KEY (`char_id`,`key`,`index`),
   KEY `char_id` (`char_id`)
 ) ENGINE=MyISAM;
@@ -403,7 +403,7 @@ CREATE TABLE IF NOT EXISTS `global_acc_reg_num` (
   `account_id` int(11) unsigned NOT NULL default '0',
   `key` varchar(32) binary NOT NULL default '',
   `index` int(11) unsigned NOT NULL default '0',
-  `value` int(11) NOT NULL default '0',
+  `value` bigint(11) NOT NULL default '0',
   PRIMARY KEY (`account_id`,`key`,`index`),
   KEY `account_id` (`account_id`)
 ) ENGINE=MyISAM;
